running corresponding route handlers, public router interfaces moved to separate file
